adapt documents schema to match relay specs

+++ services/document/documentService.ts
@@ -4,6 +4,7 @@ import { getDatabase } from "../dbService.ts";
 import { NewDocument } from "./types.ts";
 import { DocumentWithAccesLevel } from "./types.ts";
 import { AccessLevel } from "./types.ts";
+import { getUserGroups } from "../group/groupService.ts";
 
 const db = getDatabase();
 
@@ -33,13 +34,19 @@ export async function getUserDocuments(
     }
   }
 
-  return userDocuments.map((
+  const docsWithAccessLevels = userDocuments.map((
     d: Document,
   ) => ({
     ...d,
     accessLevel: d.permissions[userId] ||
-      (d.userId === userId ? AccessLevel.Manage : AccessLevel.None),
+      (d.userId === userId ? AccessLevel.Delete : AccessLevel.None),
   }));
+
+  docsWithAccessLevels.sort((d1, d2) =>
+    new Date(d2.createdAt).getTime() - new Date(d1.createdAt).getTime()
+  );
+
+  return docsWithAccessLevels;
 }
 
 export async function getGroupDocuments(
@@ -63,7 +70,7 @@ export async function addDocument({
   content,
   userId,
   type,
-}: NewDocument): Promise<Document | null> {
+}: NewDocument): Promise<DocumentWithAccesLevel | null> {
   const id = crypto.randomUUID();
   await db.set([documentsPrefix, id], {
     id,
@@ -75,17 +82,42 @@ export async function addDocument({
     createdAt: new Date().toISOString(),
   });
 
-  return getDocumentById(id);
+  const newDocument = await getDocumentById(id);
+
+  return newDocument && { ...newDocument, accessLevel: AccessLevel.Delete };
+}
+
+async function getDocumentAccessLevel(
+  document: Document,
+  userId: string,
+): Promise<AccessLevel> {
+  const userAccessLvl: AccessLevel = userId === document.userId
+    ? AccessLevel.Delete
+    : document.permissions[userId] || AccessLevel.None;
+
+  const userGroups = await getUserGroups(userId);
+  const groupAccessLvl = Math.max(
+    ...userGroups.map((g) => document?.permissions[g.id] || AccessLevel.None),
+  );
+
+  return Math.max(userAccessLvl, groupAccessLvl);
 }
 
 export async function updateDocument(
   id: string,
-  { type, title, content }: Omit<NewDocument, "userId">,
-): Promise<Document | null> {
+  {
+    type,
+    title,
+    content,
+    contributorId,
+  }: Omit<NewDocument, "userId"> & { contributorId: string },
+): Promise<DocumentWithAccesLevel | null> {
   let doc: Document | null = await getDocumentById(id);
+  let accessLevel: AccessLevel = AccessLevel.None;
 
   if (doc) {
-    await db.set([documentsPrefix, id], {
+    accessLevel = await getDocumentAccessLevel(doc, contributorId);
+    accessLevel > AccessLevel.View && await db.set([documentsPrefix, id], {
       id,
       type: type || doc.type,
       title: title || doc.title,
@@ -99,17 +131,22 @@ export async function updateDocument(
     doc = await getDocumentById(id);
   }
 
-  return doc;
+  return doc && { ...doc, accessLevel };
 }
 
 export async function updateDocumentPermissions(
   id: string,
-  permissions: Permission[],
-): Promise<Document | null> {
-  const doc: Document | null = await getDocumentById(id);
+  { permissions, contributorId }: {
+    permissions: Permission[];
+    contributorId: string;
+  },
+): Promise<DocumentWithAccesLevel | null> {
+  let doc: Document | null = await getDocumentById(id);
+  let accessLevel: AccessLevel = AccessLevel.None;
 
   if (doc) {
-    await db.set([documentsPrefix, id], {
+    accessLevel = await getDocumentAccessLevel(doc, contributorId);
+    accessLevel > AccessLevel.Update && await db.set([documentsPrefix, id], {
       ...doc,
       permissions: {
         ...doc.permissions,
@@ -119,13 +156,24 @@ export async function updateDocumentPermissions(
       },
       updatedAt: new Date().toISOString(),
     });
+
+    doc = await getDocumentById(id);
   }
 
-  return getDocumentById(id);
+  return doc && { ...doc, accessLevel };
 }
 
-export async function deleteDocument(id: string): Promise<Document | null> {
-  await db.delete([documentsPrefix, id]);
+export async function deleteDocument(
+  id: string,
+  userId: string,
+): Promise<Document | null> {
+  const doc: Document | null = await getDocumentById(id);
+
+  if (doc) {
+    const accessLevel = await getDocumentAccessLevel(doc, userId);
+    accessLevel === AccessLevel.Delete &&
+      await db.delete([documentsPrefix, id]);
+  }
 
   return getDocumentById(id);
 }
